backend/agrox-api/src/dev-scripts/python-area/pyproject.toml
[tool.poetry]
name = "python-area"
version = "0.1.0"
description = ""
authors = ["hjuste <hugo.juste@embraer.com.br>"]

[tool.poetry.dependencies]
python = "^3.7"
openpyxl = "^3.0.5"
python-dateutil = "^2.8.1"
pillow = "^7.2.0"
pandas = "^1.1.2"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
